Atualização dos gráficos para uso de dados externos
