+++ database/financeiro.json
@@ -1,0 +1,16 @@
+[{
+    "labels":["Receita", "Despesas", "Lucros"],
+    "data":[200, 109, 30]
+},
+{    
+    "labels": ["Serviços", "Produtos", "Lipeza","Pessoal","Manutenção"],
+    "data": [400, 309, 130,358,400]
+},
+{
+    "labels": ["Janeiro", "Fevereiro", "Março", "Abril", "Maio"],
+    "data": [100, 500, 250, 300, 90]
+},
+{
+    "labels": ["Janeiro", "Fevereiro", "Março", "Abril", "Maio"],
+    "data": [200, 250, 300, 150, 100]
+}]
