+++ routes/financeiroRoute.js
@@ -1,15 +1,10 @@
-var express = require('express');
-var router = express.Router();
+const express = require('express');
+
+const financeiroController = require('../controllers/financeiroController');
+
+const router = express.Router();
 
 /* GET financeiro page. */
-router.get('/', function(req, res, next) {
-  let tabActive = {homeAct: "inactive",
-                  adminAct: "inactive",
-                  financeiroAct: "active",
-                  clientesAct: "inactive",
-                  funcionariosAct: "inactive",
-                  pdvAct: "inactive"};
-  res.render('financeiro', { title: 'Express', tabs: tabActive });
-});
+router.get('/', financeiroController.index);
 
 module.exports = router;
