+++ views/financeiro.ejs
@@ -124,10 +124,131 @@
     </main>
 
     <%- include('./partials/footer') %>
+    <%= dataFinanceiro[0].data %>
 
     <%- include('./partials/scripts') %>
-
+    
+    <!-- Importa o objeto dataFinanceiro de forma que os scripts seguintes consigam utilizá-lo -->
+    <script>
+        var dbData = <%- JSON.stringify(dataFinanceiro) %>;
     </script>
+    
+    <script>
+        var ctx = document.getElementById('barra').getContext('2d');
+                
+        var myChart = new Chart(ctx, {
+            type: 'horizontalBar',
+            data: {
+                labels: dbData[0].labels,
+                datasets: [{
+                    label:'Receita x Despesas',
+                    data: dbData[0].data,
+                    backgroundColor: [
+                        'rgba(255, 99, 132, 0.2)',
+                        'rgba(54, 162, 235, 0.2)',
+                        'rgba(255, 206, 86, 0.2)'
+                        
+                    ],
+                    borderColor: [
+                        'rgba(255, 99, 132, 1)',
+                        'rgba(54, 162, 235, 1)',
+                        'rgba(255, 206, 86, 1)',
+                        
+                    ],
+                    borderWidth: 1
+                }]
+            },
+            options: {
+                scales: {
+                    yAxes: [{
+                        ticks: {
+                            beginAtZero: true
+                        }
+                    }]
+                }
+            }
+        });
+    </script>
+
+        <script>
+            var ctx = document.getElementById('pizza-despesa').getContext('2d');
+            var myChart = new Chart(ctx, {
+            type: 'doughnut',
+            data: {
+                labels: dbData[1].labels,
+                datasets: [{
+                    
+                    data: dbData[1].data,
+                    backgroundColor: [
+                        'rgba(255, 99, 132, 0.2)',
+                        'rgba(54, 162, 235, 0.2)',
+                        'rgba(255, 206, 86, 0.2)',
+                        'rgba(154, 62, 255, 0.2)',
+                        'rgba(100, 206, 86, 0.2)'
+                        
+                    ],
+                    borderColor: [
+                        'rgba(255, 99, 132, 1)',
+                        'rgba(54, 162, 235, 1)',
+                        'rgba(255, 206, 86, 1)',
+                        'rgba(154, 62, 255, 1)',
+                        'rgba(100, 206, 86, 1)',
+                        
+                    ],
+                    borderWidth: 1
+                }]
+            },
+            options: {
+                
+            }
+        });
+        </script>
+
+        <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
+        <script type="text/javascript">
+          google.charts.load("current", {packages:["corechart"]});
+          google.charts.setOnLoadCallback(drawChart);
+          function drawChart() {
+            var data = google.visualization.arrayToDataTable([
+              ['Task', 'Hours per Day'],
+              ['Work',     11],
+              ['Eat',      2],
+              ['Commute',  2],
+              ['Watch TV', 2],
+              ['Sleep',    7]
+            ]);
+        
+            var options = {
+              title: 'My Daily Activities',
+              pieHole: 0.4,
+            };
+        
+            var chart = new google.visualization.PieChart(document.getElementById('donutchart'));
+            chart.draw(data, options);
+          }
+        </script>
+        
+        <!-- *********************************************************************
+        Gráfico de barras com linhas****************************************** -->
+        
+        <script>
+            var ctx = document.getElementById('bar-line').getContext('2d');
+            var myChart = new Chart(ctx, {
+                type: 'bar',
+                data: {
+                    datasets: [{
+                        label: 'Gráfico de barras',
+                        data: dbData[2].data
+                    },
+                {
+                    label: 'Gráfico de linha',
+                    data: dbData[3].data,
+                    type: 'line'
+                }],
+                labels: dbData[2].labels
+                } 
+            });
+            </script>
 </body>
 
 </html>
