+++ views/partials/scripts.ejs
@@ -1,116 +1,1 @@
 <script src="https://cdn.jsdelivr.net/npm/chart.js@2.8.0"></script>
-<script>
-   
-var ctx = document.getElementById('barra').getContext('2d');
-var myChart = new Chart(ctx, {
-    type: 'horizontalBar',
-    data: {
-        labels: ['Receita', 'Despesas', 'Lucros'],
-        datasets: [{
-            label:'Receita x Despesas',
-            data: [200, 109, 30],
-            backgroundColor: [
-                'rgba(255, 99, 132, 0.2)',
-                'rgba(54, 162, 235, 0.2)',
-                'rgba(255, 206, 86, 0.2)'
-                
-            ],
-            borderColor: [
-                'rgba(255, 99, 132, 1)',
-                'rgba(54, 162, 235, 1)',
-                'rgba(255, 206, 86, 1)',
-                
-            ],
-            borderWidth: 1
-        }]
-    },
-    options: {
-        scales: {
-            yAxes: [{
-                ticks: {
-                    beginAtZero: true
-                }
-            }]
-        }
-    }
-});
-</script>
-<script>
-    var ctx = document.getElementById('pizza-despesa').getContext('2d');
-var myChart = new Chart(ctx, {
-    type: 'doughnut',
-    data: {
-        labels: ['Serviços', 'Produtos', 'Lipeza','Pessoal','Manutenção'],
-        datasets: [{
-            
-            data: [400, 309, 130,358,400],
-            backgroundColor: [
-                'rgba(255, 99, 132, 0.2)',
-                'rgba(54, 162, 235, 0.2)',
-                'rgba(255, 206, 86, 0.2)',
-                'rgba(154, 62, 255, 0.2)',
-                'rgba(100, 206, 86, 0.2)'
-                
-            ],
-            borderColor: [
-                'rgba(255, 99, 132, 1)',
-                'rgba(54, 162, 235, 1)',
-                'rgba(255, 206, 86, 1)',
-                'rgba(154, 62, 255, 1)',
-                'rgba(100, 206, 86, 1)',
-                
-            ],
-            borderWidth: 1
-        }]
-    },
-    options: {
-        
-    }
-});
-</script>
-<script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
-<script type="text/javascript">
-  google.charts.load("current", {packages:["corechart"]});
-  google.charts.setOnLoadCallback(drawChart);
-  function drawChart() {
-    var data = google.visualization.arrayToDataTable([
-      ['Task', 'Hours per Day'],
-      ['Work',     11],
-      ['Eat',      2],
-      ['Commute',  2],
-      ['Watch TV', 2],
-      ['Sleep',    7]
-    ]);
-
-    var options = {
-      title: 'My Daily Activities',
-      pieHole: 0.4,
-    };
-
-    var chart = new google.visualization.PieChart(document.getElementById('donutchart'));
-    chart.draw(data, options);
-  }
-</script>
-
-<!-- *********************************************************************
-Gráfico de barras com linhas****************************************** -->
-
-<script>
-   
-    var ctx = document.getElementById('bar-line').getContext('2d');
-    var myChart = new Chart(ctx, {
-        type: 'bar',
-        data: {
-            datasets: [{
-                label: 'Gráfico de barras',
-                data: [100, 500, 250, 300, 90]
-            },
-        {
-            label: 'Gráfico de linha',
-            data: [200, 250, 300, 150, 100],
-            type: 'line'
-        }],
-        labels: ['Janeiro', 'Fevereiro', 'Março', 'Abril', 'Maio']
-        } 
-    });
-    </script>
